+++ src/functions.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once 'PHPHelperId/ribuan.php';
+require_once 'PHPHelperId/re_ribuan.php';
 require_once 'PHPHelperId/romawi.php';
 require_once 'PHPHelperId/rupiah.php';
 require_once 'PHPHelperId/terbilang.php';